Get environment-variables for proxy

envvar_proxy()

returns first value found in the sequence of environment variables: "http_proxy", "HTTP_PROXY", "https_proxy", "HTTPS_PROXY".

envvar_no_proxy()

returns first value found in the sequence of environment variables: "no_proxy", "NO_PROXY".

envvar_proxy()

envvar_no_proxy()

Value

character value of proxy or no_proxy environment-variable.

Details

Note that envvar_proxy() returns "" for an unset environment-variable. This is different from get_proxy(), which will return NA if the proxy should not be used. This reflects the difference in syntax between between Sys.getenv() and withr::with_envvar() .

Examples

  envvar_proxy()

#> [1] ""

  envvar_no_proxy()

#> [1] ""