Set proxy environment-variables temporarily

Given a url and the your current proxy and no-proxy environment-variables, these functions temporarily set your proxy environment-variables:

local_proxy_for_url()

Set until the current evaluation context ends (such as the end of a function).

with_proxy_for_url()

Set only for the execution of the enclosed code.

without_proxy()

For the execution of the enclosed code, regardless of url, unsets the proxy environment-variables.

These functions local_proxy_for_url() and with_proxy_for_url() are useful in situations where the proxy variables are effective, and the no-proxy variable is not effective.

The function without_proxy() particularly useful for when you need to call another function interactively, but need to turn the proxy off.

local_proxy_for_url(url, .local_envir = parent.frame())

with_proxy_for_url(url, code)

without_proxy(code)

Arguments

url
character address for which to determine the proxy.

.local_envir
environment, the environment to use for scoping.

code
The code to execute in the temporary environment.

Value

These functions are called for their side-effects:

local_proxy_for_url()

character vector of previous values of proxy environment-variables, invisibly.

with_proxy_for_url()

The results of the evaluation of the code argument.

without_proxy()

The results of the evaluation of the code argument.

Details

These functions themselves call withr::withr with_ and local_ functions; they behave in the same way as the withr functions.

Examples

# sets your proxy variables to the right value for a given `url`
local_proxy_for_url("https://github.acme.com")
envvar_proxy()

#> [1] ""

withr::deferred_clear() # reset state

# same as above, but for enclosed code
with_proxy_for_url(
  "https://github.acme.com",
  envvar_proxy()
)

#> [1] ""

# run with proxy off
without_proxy(
  envvar_proxy()
)

#> [1] ""