Determine proxy

Use these functions to determine if a proxy should be used for a given url and no_proxy string.

uses_proxy(url, no_proxy = envvar_no_proxy())

get_proxy(url, proxy = envvar_proxy(), no_proxy = envvar_no_proxy())

Arguments

url
character address for which to determine the proxy.

no_proxy
character comma-separated string containing hostnames where the proxy is not to be used.

proxy
character the proxy that would be used.

Value

uses_proxy()

logical indicating if the proxy should be used.

get_proxy()

character value of proxy to use, NA if no proxy.

Details

This aims to emulate the cURL specification for NO_PROXY. The biggest difference between this implementation and the cURL specification is that the wildcard character (*) is used more permissibly here.

Note that get_proxy() returns NA if the proxy should not be used; this is the value used to set the proxy environment-variables. This is different from envvar_proxy(), which will return "" for an unset environment-variable. This reflects the difference in syntax between between withr::with_envvar() and Sys.getenv().

Examples

  proxy <- "http://proxy.acme.com"
  no_proxy <- "github.acme.com,localhost,127.0.0.1"

  uses_proxy("https://github.com", no_proxy)      # TRUE

#> [1] TRUE

  uses_proxy("https://github.acme.com", no_proxy) # FALSE

#> [1] FALSE

  get_proxy("https://github.com", proxy, no_proxy) # "http://proxy.acme.com"

#> [1] "http://proxy.acme.com"

  get_proxy("https://github.acme.com", proxy, no_proxy) # ""

#> [1] NA